Add cylinder test case to test scaling
