+++ tests/mappers/test_linear.py
@@ -1,6 +1,6 @@
 from coconut import data_structure
 from coconut.data_structure import KratosUnittest
-from coconut.tests.mappers.test_nearest import Case1D, Case2D, Case3DSphere, Case3DCylinder, Case3DSinc
+from coconut.tests.mappers.test_interpolator import Case1D, Case2D, Case3DSphere, Case3DCylinder, Case3DSinc
 from coconut.coupling_components.mappers.linear import *
 
 import os
@@ -62,18 +62,24 @@ class TestMapperLinear(KratosUnittest.TestCase):
 
         # 3D case: cylinder + sine function
         """
-        TODO
+        n_x_from, n_theta_from = 28, 56
+        n_x_to, n_theta_to = 35, 60
+            length = 10.
+            => max error = 0.021
         """
-        n_x_from, n_theta_from = 140 * 2, 28 * 2
-        n_x_to, n_theta_to = 350, 60
-        length = 106. / 7.
+        n_x_from, n_theta_from = 28, 56
+        n_x_to, n_theta_to = 35, 60
+        length = 10.
         par_mapper['settings'].SetArray('directions', ['X', 'Y', 'Z'])
+        par_mapper['settings'].AddEmptyValue('scaling')
+        par_mapper['settings'].SetArray('scaling', [.1, 1, 1])  # bad result without scaling
 
         case = Case3DCylinder(n_x_from, n_theta_from, n_x_to, n_theta_to, length)
         case.map(par_mapper)
-        # self.assertTrue(case.check(tolerance=5e-4))
-        # if gui:
-        case.plot()
+        self.assertTrue(case.check(tolerance=0.03))
+        if gui:
+            case.plot()
+        par_mapper['settings'].RemoveValue('scaling')
 
         # 3D case: sinc + linear vector function
         """
